Rename functions which ought to be named *Locked()
Several functions operate on variables to which access needs to be
synchronized.  However, it happens that the functions in question
are only ever called from places which have already synchronized.
Therefore, nothing is really wrong, but the functions ought to
have 'Locked' appended to their names, to indicate that it is the
caller's responsibility to synchronize before calling them.

Change-Id: I44e7dc0dff6da9436677cb10908dce41ffeba195